Replace __pycache__ with **/__pycache__ in .gitignore
For the Espressif fork, an 'idf.py python-clean' will remove
hw/mcu/sony/cxd56/tools/__pycache__/xmodem.cpython-36.pyc but
because it's checked in for some reason, the tinyusb submodule
is now marked as dirty.  With that submodule in that state,
you have to manually reset the submodule every time you change
esp-idf branches.

Although __pycache__ was already in .gitignore, that would only
match directories in the top of the working tree.  Prepending
**/ causes all __pycache__ directories to be ignored.