fix(dcd_dwc2): Support ISO IN transfer when bInterval > 1
dwc2 requires manually toggle Even/Odd bit manually for ISO IN transfer,
that's poses a problem when bInterval > 1 mainly for audio class, as the
moment the transfer is scheduled, we don't know when the host will issue
IN token (bInterval vs bRefresh schenanigans).

Linux driver use NAK interrupt to detect when the host is sending IN token
and toggle the Even/Odd bit accordingly based on the current frame number
and bInterval.

However on ST's stripped down DWC2 FS controller (e.g STM32F4, STM32F7),
NAK interrupt is not supported, even it's marked as always present in DWC2
databook. NAK interrupt is only supported on HS controller with external PHY.

Instead I schedule all ISO IN transfer for next frame, if the transfer failed,
incomplete isochronous IN transfer interrupt will be triggered and we can
relaunch the transfer.

This is a combination of 4 commits, including:
- dwc2: support ISO IN transfer when bInterval > 1
- Retry until bInterval
- Simply restart the transfer
- extract to handle_incomplete_iso_in() for readability

Signed-off-by: HiFiPhile <admin@hifiphile.com>